refactored call get tech records to tech record repository
