+++ src/services/CertificateGenerationService.ts
@@ -572,20 +572,20 @@ class CertificateGenerationService {
       const processRecordsRes = this.groupRecordsByStatusCode(searchResult);
 
       if (processRecordsRes.currentCount !== 0) {
-        return this.callGetTechRecords(
+        return this.techRecordsRepository.callGetTechRecords(
           processRecordsRes.currentRecords[0].systemNumber,
           processRecordsRes.currentRecords[0].createdTimestamp,
         );
       }
 
       if (processRecordsRes.provisionalCount === 1) {
-        return this.callGetTechRecords(
+        return this.techRecordsRepository.callGetTechRecords(
           processRecordsRes.provisionalRecords[0].systemNumber,
           processRecordsRes.provisionalRecords[0].createdTimestamp,
         );
       }
 
-      return this.callGetTechRecords(
+      return this.techRecordsRepository.callGetTechRecords(
         processRecordsRes.provisionalRecords[1].systemNumber,
         processRecordsRes.provisionalRecords[1].createdTimestamp,
       );
@@ -905,37 +905,6 @@ class CertificateGenerationService {
       Make: (techRecord as TechRecordType<'hgv' | 'trl'>).techRecord_make,
       Model: (techRecord as TechRecordType<'hgv' | 'trl'>).techRecord_model,
     };
-  };
-
-  /**
-   * Used to get a singular whole technical record.
-   * @param systemNumber
-   * @param createdTimestamp
-   */
-  public callGetTechRecords = async <T extends TechRecordGet['techRecord_vehicleType']>(systemNumber: string, createdTimestamp: string): Promise<TechRecordType<T> | undefined> => {
-    const config: IInvokeConfig = this.config.getInvokeConfig();
-    const invokeParams: InvocationRequest = {
-      FunctionName: config.functions.techRecords.name,
-      InvocationType: 'RequestResponse',
-      LogType: 'Tail',
-      Payload: toUint8Array(JSON.stringify({
-        httpMethod: 'GET',
-        path: `/v3/technical-records/${systemNumber}/${createdTimestamp}`,
-        pathParameters: {
-          systemNumber,
-          createdTimestamp,
-        },
-      })),
-    };
-    try {
-      const lambdaResponse = await this.lambdaClient.invoke(invokeParams);
-      const res = await this.lambdaClient.validateInvocationResponse(lambdaResponse);
-      return JSON.parse(res.body);
-    } catch (e) {
-      console.log('Error in get technical record');
-      console.log(JSON.stringify(e));
-      return undefined;
-    }
   };
 
   /**
